Fix bugs in install script
Implement 'install', which (somewhat) intelligently symlinks files in
this repository to your $HOME directory.  This is still very much a
work-in-progress.  Testing of edge cases and functionality beyond
symlinking a file that already exists in $HOME has not been tested.

This script first looks at all files in this repository.  So long as the
file does not match one of
- install       # the script itself
- $SCRIPT_PATH  # the path to the script itself
- .git/         # the .git directory
the script will symlink the file.

If there is a file in $HOME that is not a symlink and that matches the
name of the file $file that is being copied, the file is moved to
$HOME/$file.bak.  This ensures that a good copy of each file still
exists.

Furthermore, if the $HOME/$file.bak file already exists (i.e. if this is
not the first time this script has been run) then the script does NOT
overwrite this file (as it is assumed that the script is being run a
second time to recover from an error, and therefore the original backup
should not be clobbered with a possibly bad backup).  Instead, the
backup is written to a temporary directory that matches
'/tmp/install-dotfiles.?????'.

Old symlinks are removed from $HOME and a new symlink is created.

Finally, the backup file is concatenated to the end of the symlinked
file.  This ensures that original functionality from the original file
is not lost.  However, this creates a diff in the file, as the
machine-specific dotfile is now concatenated with this repository's
version.

To mitigate this, dotfiles should have logic at the end to look for the
$HOME/$file.bak file and source it as necessary so that concatenation is
not necessary.